+++ packages/extractor/src/ffmpeg-path.js
@@ -1,0 +1,28 @@
+
+const log = require('@home-gallery/logger')('extractor.ffmpeg');
+
+const { getNativeCommand } = require('./utils')
+
+const getFfmpegPaths = options => {
+  let ffmpegPath
+  if (options.useNative.includes('ffmpeg')) {
+    log.debug(`Use native system command ffmpeg`)
+    ffmpegPath = getNativeCommand('ffmpeg')
+  } else {
+    ffmpegPath = require('@ffmpeg-installer/ffmpeg').path
+  }
+
+  let ffprobePath
+  if (options.useNative.includes('ffprobe')) {
+    log.debug(`Use native system command ffprobe`)
+    ffprobePath = getNativeCommand('ffprobe')
+  } else {
+    ffprobePath = require('@ffprobe-installer/ffprobe').path
+  }
+
+  return [ffmpegPath, ffprobePath]
+}
+
+module.exports = {
+  getFfmpegPaths
+}