+++ packages/extractor/src/utils.js
@@ -1,0 +1,8 @@
+const isWindows = () => process.platform.match(/win32/i)
+
+const getNativeCommand = command => `${command}${isWindows() ? '.exe' : ''}`
+
+module.exports = {
+  isWindows,
+  getNativeCommand
+}