+++ packages/extractor/src/video-poster.js
@@ -1,12 +1,11 @@
 const log = require('@home-gallery/logger')('extractor.video.poster');
 
 const { resizeImage } = require('./image-preview');
-const { extractVideoFames } = require('./video-frames');
 const { toPipe, conditionalTask } = require('./task');
 
 const videoPosterSuffix = 'video-poster.jpg';
 
-function videoPoster(storage, previewImageSizes) {
+function videoPoster(storage, { imageResizer, videoFrameExtractor, imagePreviewSizes }) {
 
   const test = entry => entry.type === 'video' && !storage.hasEntryFile(entry, videoPosterSuffix);
 
@@ -14,14 +13,14 @@ function videoPoster(storage, previewImageSizes) {
     const t0 = Date.now();
     const dir = storage.getEntryDirname(entry);
     const basename = storage.getEntryBasename(entry, videoPosterSuffix);
-    extractVideoFames(entry.src, dir, basename, 1, (err) => {
+    videoFrameExtractor(entry.src, dir, basename, 1, (err) => {
       if (err) {
         log.warn(err, `Could not extract video frame from ${entry}: ${err}`)
         return cb();
       }
 
       const posterSrc = storage.getEntryFilename(entry, videoPosterSuffix);
-      resizeImage(storage, entry, posterSrc, previewImageSizes, (err, calculatedSizes) => {
+      resizeImage(storage, imageResizer, entry, posterSrc, imagePreviewSizes, (err, calculatedSizes) => {
         if (err) {
           log.warn(err, `Could not resize video frame from ${entry}: ${err}`)
           return cb();