+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- extractor: Option to use native system executables like vipsthumbnail or ffmpeg
 - extractor: Add support for heic/heif image format
 - extractor: Extract embedded previews from raw files
 - query: Add cmp, (all) in list and range to common properties